It just works

+++ Blocktrader.Exchange/Binance/Models/OrderBookResponse.cs
@@ -2,7 +2,7 @@ namespace Blocktrader.Exchange.Binance.Models
 {
     internal class OrderBookResponse
     {
-        public int LastUpdateId { get; set; }
+        public long LastUpdateId { get; set; }
         
         public int Timestamp { get; set; }
