+++ Blocktrader.Exchange/Bitstamp/Models/OrderBookResponse.cs
@@ -5,7 +5,7 @@ namespace Blocktrader.Exchange.Bitstamp.Models
     [Obsolete("Проверить на лишние поля")]
     internal class OrderBookResponse
     {
-        public int LastUpdateId { get; set; }
+        public long LastUpdateId { get; set; }
         
         public int Timestamp { get; set; }
